refactor: remove legacy sys.modules backward-compat shims
Direct imports replace the old module aliases; tests updated accordingly.
